+++ src/preview/components/Markdown.tsx
@@ -5,6 +5,8 @@ import rehypePrettyCode from 'rehype-pretty-code';
 import { usePrevious } from '../hooks/usePrevious';
 import { messageHandler } from '@estruyf/vscode/dist/client/webview';
 import { WebViewMessages } from '../../constants';
+import { renderToString } from 'react-dom/server';
+import { convertTemplateToHtml } from '../../utils/convertTemplateToHtml';
 
 export interface IMarkdownProps {
   content?: string;
@@ -24,7 +26,10 @@ export const Markdown: React.FunctionComponent<IMarkdownProps> = ({
   updateBgStyles
 }: React.PropsWithChildren<IMarkdownProps>) => {
   const prevContent = usePrevious(content);
+  const [isReady, setIsReady] = React.useState(false);
   const [customTheme, setCustomTheme] = React.useState<string | undefined>(undefined);
+  const [customLayout, setCustomLayout] = React.useState<string | undefined>(undefined);
+  const [template, setTemplate] = React.useState<string | undefined>(undefined);
 
   const {
     markdown,
@@ -48,7 +53,30 @@ export const Markdown: React.FunctionComponent<IMarkdownProps> = ({
     }
   });
 
+  const prevMatter = usePrevious(JSON.stringify(matter));
 
+  const updateCustomLayout = React.useCallback((layout: string, metadata: any) => {
+    if (layout) {
+      messageHandler.request<string>(WebViewMessages.toVscode.getFileContents, layout).then(async (templateHtml) => {
+        if (templateHtml) {
+          let html = await convertTemplateToHtml(templateHtml, {
+            metadata,
+            content: renderToString(markdown),
+          });
+
+          // Replace all the `<style>` tags with `<style type="text/tailwindcss">`
+          html = html.replace(/<style>/g, '<style type="text/tailwindcss">');
+
+          setTemplate(html);
+          setIsReady(true);
+        }
+      }).catch(() => {
+        setIsReady(true);
+      });
+    } else {
+      setIsReady(true);
+    }
+  }, [content, markdown]);
 
   const updateCustomThemePath = React.useCallback((customThemePath: string) => {
     if (!customThemePath) {
@@ -68,8 +96,25 @@ export const Markdown: React.FunctionComponent<IMarkdownProps> = ({
   }, []);
 
   React.useEffect(() => {
+    if (!matter) {
+      setIsReady(false);
+      return;
+    }
+
+    if (prevMatter === JSON.stringify(matter)) {
+      return;
+    }
+
+    setIsReady(false);
+    setCustomLayout(undefined);
+    setCustomTheme(undefined);
+    setTemplate(undefined);
+
+    const cLayout = matter?.customLayout || undefined;
+    updateCustomLayout(cLayout, matter);
+
     updateTheme(matter?.theme || "default");
-    updateLayout(matter?.layout || "default");
+    updateLayout(cLayout || matter?.layout || "default");
     updateCustomThemePath(matter?.customTheme || undefined);
 
     if (matter?.image) {
@@ -84,7 +129,7 @@ export const Markdown: React.FunctionComponent<IMarkdownProps> = ({
     } else {
       updateBgStyles(undefined);
     }
-  }, [matter, updateTheme, updateLayout, updateCustomThemePath, updateBgStyles, webviewUrl]);
+  }, [isReady, prevContent, matter, updateTheme, updateLayout, updateCustomThemePath, updateBgStyles, webviewUrl]);
 
   React.useEffect(() => {
     if (content && content !== prevContent) {
@@ -93,11 +138,29 @@ export const Markdown: React.FunctionComponent<IMarkdownProps> = ({
     }
   }, [content, vsCodeTheme]);
 
+  if (!isReady) {
+    return null;
+  }
+
+  if (customLayout && !template) {
+    return null;
+  }
+
+  if (template) {
+    return (
+      <>
+        {customTheme && <link href={customTheme} rel="stylesheet" />}
+
+        <div className='slide__content__custom' dangerouslySetInnerHTML={{ __html: template }} />
+      </>
+    );
+  }
+
   return (
     <>
       {customTheme && <link href={customTheme} rel="stylesheet" />}
 
-      {markdown}
+      <div className='slide__content__inner'>{markdown}</div>
     </>
   );
 };