+++ src/utils/convertTemplateToHtml.ts
@@ -1,0 +1,11 @@
+import Handlebars from "handlebars";
+
+export const convertTemplateToHtml = async (template: string, data: any) => {
+  Handlebars.registerHelper("eq", (a, b) => {
+    return a === b;
+  });
+
+  const templateFunction = Handlebars.compile(template);
+  const html = templateFunction(data);
+  return html;
+};