+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Change Log
 
+## [0.0.87] - 2025-03-xx
+
+- [#74](https://github.com/estruyf/vscode-demo-time/issues/74): Support for custom slide layouts
+- [#75](https://github.com/estruyf/vscode-demo-time/issues/75): Provide a filename for the PDF export
+
 ## [0.0.86] - 2025-03-20
 
 - Optimized package size